[refactor/#224] modal 폴더명 PascalCase로 변경, import 경로 수정
