+++ src/components/ActivityReservation/ActivityReservationBar.tsx
@@ -1,5 +1,5 @@
 import Button from "@/src/components/Button/Button";
-import ReservationModal from "@/src/components/modal/ReservationModal/ReservationModal";
+import ReservationModal from "@/src/components/Modal/ReservationModal/ReservationModal";
 import { useCalendar } from "@/src/stores/CalendarStore";
 import useModalStore from "@/src/stores/ModalStore";
 import { ActivityDetailResponse } from "@/src/types/activitiesResponses";
